requirements.txt
absl-py==0.7.0
ftfy
gym==0.17.3
hydra-core==1.0.5
imageio==2.9.0
imageio-ffmpeg==0.4.5
kornia==0.4.1
matplotlib==3.1.1
meshcat==0.0.18
numpy==1.18.5
opencv-python==4.1.2.30
pybullet==3.0.4
pytorch_lightning==1.0.3
regex
scikit-image==0.17.2
scipy==1.4.1
tdqm
torch==1.7.1
torchvision==0.8.2
transformers==4.3.2
transforms3d==0.3.1
wandb==0.10.15